chore(page) refactored Shell.tsx code
