+++ src/Shell/Shell.tsx
@@ -2,24 +2,22 @@
  * Copyright (C) Verizon. All rights reserved.
  */
 
-import React, { useEffect, useRef, useState } from "react";
+import React, { useEffect } from "react";
 import { Outlet } from "@tanstack/react-router";
 import { useNavigate } from "@tanstack/react-router";
 import { useUserContext } from "@/context/UserContext";
-import { getInitials } from "@/lib/utils";
+import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { useShellStyles } from "./Shell.styles";
+import { Header } from "@/components/Header";
+import { Footer } from "@/components/Footer";
 import type { ShellProps } from "./Shell.types";
 
-/**
- * Render the final JSX of Shell
- */
+const queryClient = new QueryClient();
+
 export const Shell: React.FC<ShellProps> = () => {
   const styles = useShellStyles();
+  const { isAuthenticated } = useUserContext();
   const navigate = useNavigate();
-  const dropdownRef = useRef<HTMLDivElement>(null);
-
-  const [isDropdownOpen, setIsDropdownOpen] = useState(false);
-  const { userDetails, isAuthenticated, logout } = useUserContext();
 
   useEffect(() => {
     if (!isAuthenticated) {
@@ -27,102 +25,15 @@ export const Shell: React.FC<ShellProps> = () => {
     }
   }, [isAuthenticated, navigate]);
 
-  useEffect(() => {
-    const handleClickOutside = (event: MouseEvent) => {
-      if (
-        dropdownRef.current &&
-        !dropdownRef.current.contains(event.target as Node)
-      ) {
-        setIsDropdownOpen(false);
-      }
-    };
-
-    document.addEventListener("mousedown", handleClickOutside);
-    return () => {
-      document.removeEventListener("mousedown", handleClickOutside);
-    };
-  }, []);
-
   return (
-    <div className={styles.root}>
-      {/* Header */}
-      <header className={styles.header}>
-        <div className={styles.container}>
-          <div className="flex items-center">
-            <a href="/" className={styles.logo}>
-              Verizon
-            </a>
-            <nav className={styles.nav}>
-              {/* Uncomment possible nav links */}
-              {/* <a href="#">Mobile</a>
-              <a href="#">Internet</a>
-              <a href="#">Solutions</a>
-              <a href="#">Insights</a> */}
-            </nav>
-          </div>
-          <div className={styles.userSection}>
-            {userDetails && (
-              <span className={styles.userName}>
-                Welcome, {userDetails.name}
-              </span>
-            )}
-            <div
-              className={styles.dropdownContainer}
-              ref={dropdownRef}
-            >
-              <button
-                className={styles.searchIcon}
-                onClick={() => setIsDropdownOpen(!isDropdownOpen)}
-              >
-                {getInitials(userDetails?.name)}
-              </button>
-              {isDropdownOpen && (
-                <div className={styles.dropdown}>
-                  <a
-                    className={styles.dropdownLink}
-                    href="/"
-                  >
-                    Home
-                  </a>
-                  <a
-                    className={styles.dropdownLink}
-                    href="/submission-detail"
-                  >
-                    Submission Detail
-                  </a>
-                  <button
-                    className={styles.dropdownLink}
-                    onClick={() => {
-                      setIsDropdownOpen(false);
-                      logout();
-                    }}
-                  >
-                    Logout
-                  </button>
-                </div>
-              )}
-            </div>
-          </div>
+    <QueryClientProvider client={queryClient}>
+        <div className={styles.root}>
+          <Header />
+          <main className={styles.main}>
+            <Outlet /> {/* Child routes will render here */}
+          </main>
+          <Footer />
         </div>
-      </header>
-
-      {/* Portal Page Content */}
-      <main className={styles.main}>
-        <Outlet /> {/* Child routes will render here */}
-      </main>
-
-      {/* Footer */}
-      <footer className={styles.footer}>
-        <div>
-          <a href="#" className={styles.footerLink}>
-            Privacy Policy
-          </a>
-          <a href="#" className={styles.footerLink}>
-            Terms of Service
-          </a>
-          <a href="#">Contact Us</a>
-        </div>
-      </footer>
-    </div>
+    </QueryClientProvider>
   );
 };
