+++ src/components/Footer/Footer.styles.ts
@@ -1,0 +1,20 @@
+/*!
+ * Copyright (C) Verizon. All rights reserved.
+ */
+
+import { createUseStyles } from 'react-jss';
+
+/**
+ * Styles for the Footer slots
+ */
+export const useFooterStyles = createUseStyles({
+  footer: {
+    backgroundColor: '#f1f1f1',
+    padding: '1rem 1.5rem',
+    marginTop: 'auto',
+    textAlign: 'center',
+  },
+  footerLink: {
+    marginRight: '2rem',    
+  }
+});
