+++ src/components/Footer/Footer.types.ts
@@ -1,0 +1,9 @@
+/*!
+ * Copyright (C) Verizon. All rights reserved.
+ */
+
+/**
+ *  FooterProps
+ */
+export type FooterProps = React.PropsWithChildren & {
+};
