+++ src/components/Footer/index.ts
@@ -1,0 +1,7 @@
+/*!
+ * Copyright (C) Verizon. All rights reserved.
+ */
+
+export * from './Footer';
+export * from './Footer.types';
+export * from './Footer.styles';
