+++ src/components/Header/Header.types.ts
@@ -1,0 +1,9 @@
+/*!
+ * Copyright (C) Verizon. All rights reserved.
+ */
+
+/**
+ *  HeaderProps
+ */
+export type HeaderProps = React.PropsWithChildren & {
+};
