+++ src/components/Header/index.ts
@@ -1,0 +1,7 @@
+/*!
+ * Copyright (C) Verizon. All rights reserved.
+ */
+
+export * from './Header';
+export * from './Header.types';
+export * from './Header.styles';
