+++ src/components/Header/stories/Header.stories.tsx
@@ -1,0 +1,28 @@
+/*!
+ * Copyright (C) Verizon. All rights reserved.
+ */
+
+import * as React from 'react';
+import { Meta, StoryFn } from '@storybook/react';
+import { withPerformance } from 'storybook-addon-performance';
+
+import { Header, HeaderProps } from './..';
+
+export default {
+  title: 'Controls/Header',
+  component: Header,
+  decorators: [withPerformance],
+} as Meta;
+
+const Template: StoryFn<HeaderProps> = (args) => {
+  return <Header {...args} />;
+};
+
+export const Default = Template.bind({});
+
+Default.args = {} as HeaderProps;
+
+Default.parameters = {
+  options: { fluentVersion: 9 },
+  controls: { disable: false },
+};
