+++ src/main.tsx
@@ -7,9 +7,8 @@ import ReactDOM from "react-dom/client";
 
 import "./index.css";
 import "./tailwind.css";
+import { UserProvider } from "@/context/UserContext";
 import { RouterProvider } from "@tanstack/react-router";
-import { UserProvider } from "./context/UserContext";
-import { QueryClient, QueryClientProvider } from "@tanstack/react-query";
 import { router } from "@/routes"; // Import the router created in Shell.tsx
 
 if (process.env.NODE_ENV === "development") {
@@ -19,19 +18,14 @@ if (process.env.NODE_ENV === "development") {
   });
 }
 
-const queryClient = new QueryClient();
-
 const root = ReactDOM.createRoot(
   document.getElementById("root") as HTMLElement
 );
 
 root.render(
   <React.StrictMode>
-      <QueryClientProvider client={queryClient}>
-        <UserProvider>
-          <RouterProvider router={router} />
-        </UserProvider>
-      </QueryClientProvider>
+    <UserProvider>
+      <RouterProvider router={router} />
+    </UserProvider>
   </React.StrictMode>
 );
-
